token_service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-jose[cryptography]==3.3.0
werkzeug>=3.0.0
requests==2.31.0
PyJWT==2.8.0  # Wichtig: PyJWT hinzufügen

